The Chromosome was using the length of the given BitSet to determine its length. However, the BitSet.length() method returns the position+1 of the highest _set_ bit in the set, not the highest bit actually manipulated. The BitSet.size() method returns the number of bits allocated to the set, which is usually higher than the number of bits actually manipulated.
Since BitSet.length was being used, the chromosome lengths were getting
all screwy resulting in a loss of activity in chromosomes housing BitSets
with a small "logical" length. Furthermore, the possibility for zero-length
chromosomes existed, which caused the random number generated to get
unhappy and throw an exception.

To fix the above, I added a new parameter--chromosoneLength--to the
constructor. This value is now used internally rather than the length
of the BitSet. I modified the factory methods and the clone method
appropriately.

Finally, the clone method wasn't propegating its mutation factor
correctly, which I fixed as well.


git-svn-id: file:///home/mast/dev/jgap/svn/trunk@4 3924feb9-9058-4ba4-b128-58a6af0bd8c6